# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Build Your Christmas API
  description: |
    REST API for **Build Your Christmas** — single-brand Christmas tree e-commerce platform
    với 3 sizes (S/M/L) × 6 styles × ~25 accessories, Christmas tree editor (live quote),
    HCM-only delivery, design save/share, và admin catalog management.

    Backend is the source of truth for:
    - authentication & user identity (single-brand: chỉ roles `buyer` và `admin`)
    - catalog (Tree, Style, Accessory)
    - tree design editor (TreeDesign with live pricing)
    - shopping cart (DesignConfig-based, không phải product-based)
    - order lifecycle (HCM-only delivery, snapshot design, atomic stock reservation)
    - payment (mock, COD + online)
    - shipment tracking (HCM Delivery in-house)
    - admin management (catalog CRUD + orders + stats + users)
  version: 2.0.0
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.buildyourchristmas.vn
    description: Production (TBD)
...
paths:
...
  /api/catalog/quote:
    post:
      tags: [Catalog]
      summary: Live price preview from DesignConfig (no auth)
      operationId: quoteDesign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config: { $ref: '#/components/schemas/DesignConfig' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pricing: { $ref: '#/components/schemas/PriceBreakdown' }
        '400': { $ref: '#/components/responses/BadRequest' }

  # ============================================================
  # DESIGNS
  # ============================================================
...
  /api/designs:
    post:
      tags: [Designs]
      summary: Save a TreeDesign for the authenticated user
      operationId: createDesign
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, config]
              properties:
                name: { type: string }
                config: { $ref: '#/components/schemas/DesignConfig' }
                previewImage: { type: string }
                isPublic: { type: boolean, default: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  design: { $ref: '#/components/schemas/Design' }
                  shareUrl: { type: string, example: "/tree/my-design" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [Designs]
      summary: List my designs
      operationId: getMyDesigns
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  designs:
                    type: array
                    items: { $ref: '#/components/schemas/Design' }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/designs/{id}:
    get:
      tags: [Designs]
      summary: Get a design by Mongo _id (owner/admin only)
      operationId: getDesignById
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  design: { $ref: '#/components/schemas/Design' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Designs]
      summary: Update a design (owner only)
      operationId: updateDesign
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                config: { $ref: '#/components/schemas/DesignConfig' }
                previewImage: { type: string }
                isPublic: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  design: { $ref: '#/components/schemas/Design' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Designs]
      summary: Delete a design (owner only — preset cannot be deleted)
      operationId: deleteDesign
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/designs/{id}/duplicate:
    post:
      tags: [Designs]
      summary: Duplicate a design (own / public / preset)
      operationId: duplicateDesign
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  design: { $ref: '#/components/schemas/Design' }
                  shareUrl: { type: string }

  # ============================================================
  # CART
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Unauthorized — token missing or invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...
    DesignConfig:
      type: object
      required: [treeId, styleId, accessories, deliveryOption]
      properties:
        treeId: { type: string, description: 'Mongo ObjectId of Tree' }
        styleId: { type: string, description: 'Mongo ObjectId of Style' }
        accessories:
          type: array
          items:
            type: object
            required: [accessoryId, quantity]
            properties:
              accessoryId: { type: string, description: 'Mongo ObjectId of Accessory' }
              quantity: { type: integer, minimum: 1 }
              personalizationText: { type: string, description: 'Required if accessory.isPersonalizable. Regex ^[A-Za-z0-9 ]{1,20}$' }
        deliveryOption: { type: string, enum: [READY_TO_DISPLAY, DIY_KIT, SEPARATE] }
    PriceBreakdown:
      type: object
      properties:
        tree:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            size: { type: string, enum: [S, M, L] }
            unitPrice: { type: integer, description: 'VND' }
        style:
          type: object
          properties:
            id: { type: string }
            code: { type: string, enum: [CLASSIC, MINIMAL, GINGERBREAD, WINTER, CUTE, LUXURY] }
            name: { type: string }
        lines:
          type: array
          items:
            type: object
            properties:
              accessoryId: { type: string }
              name: { type: string }
              type: { type: string }
              group: { type: string, enum: [LIGHTS, ORNAMENT, DECOR, PERSONAL] }
              image: { type: string }
              unitPrice: { type: integer, description: 'VND' }
              quantity: { type: integer }
              lineTotal: { type: integer, description: 'VND' }
              isPersonalizable: { type: boolean }
              personalizationMaxLength: { type: integer }
              productionDays: { type: integer }
        decorationFee: { type: integer, description: 'VND — 0 unless READY_TO_DISPLAY' }
        shippingFee: { type: integer, description: 'VND — 30000' }
        unitTotal: { type: integer, description: 'VND — full price for 1 set' }
        productionDays: { type: integer }
        hasPersonalization: { type: boolean }
        hasService: { type: boolean, description: 'true if decoration fee > 0' }
        warnings:
          type: array
          items: { type: string }
    Design:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        slug: { type: string }
        year: { type: integer }
        config: { $ref: '#/components/schemas/DesignConfig' }
        isPublic: { type: boolean }
        isPreset: { type: boolean }
        duplicatedFrom: { type: string, nullable: true }
        previewImage: { type: string }
        ownerId: { type: string, nullable: true }
        shareUrl: { type: string }
        pricing: { $ref: '#/components/schemas/PriceBreakdown' }
...